prepare results for decoupling
