+++ true_north/__init__.py
@@ -1,16 +1,13 @@
 """Beautiful and pythonic benchmarks engine.
 """
-from ._check import Check
 from ._colors import Colors
 from ._group import Group
-from ._result import Result
+from . import types
 
 
 __version__ = '0.2.0'
 __all__ = [
-    'setup',
-    'Check',
     'Colors',
     'Group',
-    'Result',
+    'types',
 ]
