+++ true_north/_results/__init__.py
@@ -1,0 +1,4 @@
+
+from ._timing import TimingResult
+
+__all__ = ['TimingResult']
